WEEK 6. Improve users
