+++ src/users/dto/create-user.dto.ts
@@ -1,4 +1,4 @@
-import {IsNotEmpty, IsString} from 'class-validator';
+import { IsNotEmpty, IsString } from 'class-validator';
 
 export class CreateUserDto {
     @IsNotEmpty()
