+++ src/users/interfaces/user.interface.ts
@@ -1,8 +1,8 @@
 export interface User {
-    id: string; // uuid v4
+    id: string;
     login: string;
     password: string;
-    version: number; // integer number, increments on update
-    createdAt: number; // timestamp of creation
-    updatedAt: number; // timestamp of last update
+    version: number;
+    createdAt: number;
+    updatedAt: number;
 }
